requirements.txt
openai>=2.28.0
python-docx>=1.1.2
openpyxl>=3.1.2
xlrd>=2.0.1
rich>=13.7.0
